refactor : 노선 관리 기능 코드 리팩토링
- views 폴더로 lineManagerPage 파일 이동
- 메소드 순서 수정
