+++ src/components/views/lineManagerPage.js
@@ -2,11 +2,11 @@ import {
   LINE_MANAGER_PAGE_TEMPLATE,
   ALL_STATION_OPTION_LIST,
   LINE_TABLE_TEMPLATE,
-} from './template/lineManagerTemplate.js';
-import { lineNameValidator, lineStationsValidator } from '../utils/validator/lineValidator.js';
-import stationStorage from '../utils/stationStorage.js';
-import lineStorage from '../utils/lineStorage.js';
-import Line from '../utils/Line.js';
+} from '../template/lineManagerTemplate.js';
+import { lineNameValidator, lineStationsValidator } from '../../utils/validator/lineValidator.js';
+import stationStorage from '../../storage/stationStorage.js';
+import lineStorage from '../../storage/lineStorage.js';
+import Line from '../../utils/Line.js';
 
 export default function stationManagerPage($element) {
   $element.innerHTML = LINE_MANAGER_PAGE_TEMPLATE;
@@ -17,7 +17,7 @@ export default function stationManagerPage($element) {
   const $lineTableBody = $element.querySelector('.line-table-tbody');
 
   const stations = stationStorage().getStations();
-  let lines = lineStorage().getLine();
+  let lines = lineStorage().getLines();
 
   $startStation.innerHTML = ALL_STATION_OPTION_LIST(stations);
   $endStation.innerHTML = ALL_STATION_OPTION_LIST(stations);
@@ -33,42 +33,15 @@ export default function stationManagerPage($element) {
     return lines[lines.length - 1].id + 1;
   };
 
-  const addLine = (newLine) => {
-    lines.push(newLine);
-    lineStorage().setLine(lines);
-    renderLines();
-  };
-
-  const getStation = (stationName) => {
-    return stations.filter((station) => station.name === stationName)[0];
-  };
-
   const getStationById = (stationIds) => {
     return stations.filter((station) => station.id === stationIds)[0];
   };
 
-  const addLineInStation = (lineId, stationName) => {
-    getStation(stationName).line.push(lineId);
-    stationStorage().setStation(stations);
-  };
-
-  const createLine = (newLineName, startStationName, endStationName) => {
-    const startStationId = stationStorage().getStationIdByName(startStationName);
-    const endStationId = stationStorage().getStationIdByName(endStationName);
-    const newLine = new Line(getNewId(), newLineName);
-    newLine.setLine(startStationId, 0);
-    newLine.setLine(endStationId, 1);
-
-    addLineInStation(newLine.id, startStationName);
-    addLineInStation(newLine.id, endStationName);
-    addLine(newLine);
-  };
-
   const removeLineInStation = (currentLineId) => {
-    const stationIdsInLine = lines.filter((line) => line.id === parseInt(currentLineId))[0]
-      .stationIds;
+    const stationsInLine = lines
+      .filter((line) => line.id === parseInt(currentLineId))[0]
+      .stationIds.map(getStationById);
 
-    const stationsInLine = stationIdsInLine.map(getStationById);
     for (let i = 0; i < stationsInLine.length; i++) {
       const deleteIndex = stationsInLine[i].line.indexOf(parseInt(currentLineId));
       stationsInLine[i].line.splice(deleteIndex, 1);
@@ -92,17 +65,41 @@ export default function stationManagerPage($element) {
     }
   };
 
+  const addLine = (newLine) => {
+    lines.push(newLine);
+    lineStorage().setLine(lines);
+    renderLines();
+  };
+
+  const getStation = (stationName) => {
+    return stations.filter((station) => station.name === stationName)[0];
+  };
+
+  const addLineInStation = (lineId, stationName) => {
+    getStation(stationName).line.push(lineId);
+    stationStorage().setStation(stations);
+  };
+
+  const createLine = (newLineName, startStationName, endStationName) => {
+    const newLine = new Line(getNewId(), newLineName);
+    newLine.setLine(stationStorage().getStationIdByName(startStationName), 0);
+    newLine.setLine(stationStorage().getStationIdByName(endStationName), 1);
+
+    addLineInStation(newLine.id, startStationName);
+    addLineInStation(newLine.id, endStationName);
+    addLine(newLine);
+  };
+
   const onLineSubmitHandler = () => {
     const newLineName = $userInputLine.value;
-    const startStationName = $startStation.value;
-    const endStationName = $endStation.value;
 
     if (
       lineNameValidator(lines, newLineName) &&
-      lineStationsValidator(startStationName, endStationName)
+      lineStationsValidator($startStation.value, $endStation.value)
     ) {
-      createLine(newLineName, startStationName, endStationName);
+      createLine(newLineName, $startStation.value, $endStation.value);
     }
+
     $userInputLine.value = '';
   };
